Added ToStringOrZero() method to IMathProvider. Minor fixes.

+++ MatrixCalculator/src/Program.cs
@@ -6,7 +6,7 @@
         {
             // This generic type defines the type we will work with.
             // E. g., the type of matrix elements or the type of multiplication factor.
-            UI.UI<double>.Launch();
+            UI.UI<decimal>.Launch();
         }
     }
 }
